use modular javascript
