+++ index.html
@@ -5,6 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="styles.css">
+    <link rel="stylesheet" href="index.css">
     <title>AwsomeBooks</title>
 </head>
 <body>
@@ -51,6 +52,6 @@
     <footer>
         <div>Copyright @ CCKitur</div>
     </footer>
-<script src="main.js"></script>
+<script type="module" src="main.js"></script>
 </body>
 </html>
