+++ main.js
@@ -3,73 +3,12 @@
 /* eslint-disable max-classes-per-file */
 /* eslint-disable no-trailing-spaces */
 /* eslint-disable lines-between-class-members */
-class Book {
-  constructor(title, author) {
-    this.title = title;
-    this.author = author;
-  }
-}
-
-class UI {
-  static displayBook() {
-    const books = store.getBooks();
-    books.forEach((book) => UI.addBookToList(book));
-  }
-
-  static addBookToList(book) {
-    const list = document.querySelector('#book-list');
-    const div = document.createElement('div');
-    div.classList.add('listing');
-    div.innerHTML = `
-        <div class="book-title">"${book.title}" By </div><div class="author-div">${book.author}</div>
-        <button class='delete' id='delete'>Remove</button> 
-        `;
-    list.appendChild(div);
-  }
-
-  static deleteBook(clickTarget) {
-    if (clickTarget.classList.contains('delete')) {
-      clickTarget.parentElement.remove();
-    }
-  }
-  static clearFields() {
-    document.querySelector('#author').value = '';
-    document.querySelector('#book').value = '';
-  }
-}
-
-class store {
-  static getBooks() {
-    let books;
-    if (localStorage.getItem('books') === null) {
-      books = [];
-    } else {
-      books = JSON.parse(localStorage.getItem('books'));
-    }
-    return books;
-  }
-
-  static addBook(book) {
-    const books = store.getBooks();
-    books.push(book);
-    localStorage.setItem('books', JSON.stringify(books));
-  }
-
-  static removeBook(author) {
-    const books = store.getBooks();
-    books.forEach((book, index) => {
-      if (book.author === author) {
-        books.splice(index, 1);
-        localStorage.setItem('books', JSON.stringify(books));
-      }
-    });
-  }
-}
-
 // event  displaybook
+import UI from './modules/userInterphase.js'
+import Book from './modules/books.js'
+import store from './modules/store.js'
 
 document.addEventListener('DOMContentLoaded', UI.displayBook);
-
 document.querySelector('#booksform').addEventListener('submit', (e) => {
   e.preventDefault();
   const author = document.querySelector('#author').value;
@@ -79,6 +18,7 @@ document.querySelector('#booksform').addEventListener('submit', (e) => {
   const book = new Book(title, author);
   // add book to list
   UI.addBookToList(book);
+  
 
   store.addBook(book);
   // clearfields
