+++ styles.css
@@ -37,6 +37,8 @@ nav .right {
   border: 2px solid black;
   height: 30px;
   display: flex;
+  justify-content: center;
+  font-weight: bold;
   align-items: center;
   padding: 1% 2%;
   margin: 1% 2%;
